changed enum to constant, removed and renamed files
