+++ wls-gui-wahllokalsystem/src/types/userNotification/UserNotification.ts
@@ -1,4 +1,4 @@
-import type { UserNotificationCategoryEnum } from "@/types/wlsTypes/UserNotificationCategoryEnum.ts";
+import type { UserNotificationCategoryEnum } from "@/types/userNotification/UserNotificationCategoryEnum.ts";
 
 export interface UserNotification {
   id: string;
