+++ wls-gui-wahllokalsystem/src/types/userNotification/UserNotificationCategoryEnum.ts
@@ -1,0 +1,8 @@
+export const UserNotificationCategoryEnum = {
+  ERFOLG: "Erfolg",
+  WARNUNG: "Warnung",
+  FEHLER: "Fehler",
+} as const;
+
+export type UserNotificationCategoryEnum =
+  (typeof UserNotificationCategoryEnum)[keyof typeof UserNotificationCategoryEnum];
