--- wls-gui-wahllokalsystem/src/types/wlsTypes/UserNotificationCategoryEnum.ts
@@ -1,5 +1,0 @@
-export enum UserNotificationCategoryEnum {
-  ERFOLG = "Erfolg",
-  WARNUNG = "Warnung",
-  FEHLER = "Fehler",
-}
